requirements.txt
fastapi
motor
python-jose[cryptography]
pydantic
passlib[bcrypt]
uvicorn
pydantic-settings
